Improve SARIF round-trip and consolidate tests
This commit addresses a data loss issue when exporting and importing results
in SARIF format. Specifically, it ensures that all metadata fields, including
admin and end-user notes, are preserved by utilizing the SARIF 'properties' block.

Key changes:
- Modified `generate_sarif` to include `admin_desc` and `end-user_desc` in properties.
- Updated `import_results` to correctly map these properties and `end-user_desc`.
- Added a comprehensive round-trip test in `tests/test_sarif.py`.
- Consolidated SARIF tests by merging `tests/test_sarif_import.py` into
  `tests/test_sarif.py` and removing the redundant file.
- Verified fixes and ensured no regressions by running the full test suite.
